+++ test/datalevin/remote_dl_test.clj
@@ -474,3 +474,11 @@
            [(d/datom 1 :identifer 1) (d/datom 2 :identifer 1)]))
 
     (dc/close conn)))
+
+(deftest update-opts-test
+  (let [dir  "dtlv://datalevin:datalevin@localhost/update-opts"
+        conn (dc/create-conn dir)]
+    (is (= 100 (dc/datalog-index-cache-limit @conn)))
+    (dc/datalog-index-cache-limit @conn 0)
+    (is (= 0 (dc/datalog-index-cache-limit @conn)))
+    (dc/close conn)))